split webserver/director-v2 app module
